+++ simplesecurity/formatter.py
@@ -49,8 +49,7 @@ def formatEvidence(evidence: list[Line], newlineChar: bool = True) -> str:
 	return "\\n".join(evidenceText)
 
 
-def markdown(findings: list[Finding], heading: Optional[str] = None,
-colourMode: int = 0) -> str:
+def markdown(findings: list[Finding], heading: Optional[str] = None, colourMode: int = 0) -> str:
 	"""Format to Markdown.
 
 	Args:
@@ -63,8 +62,11 @@ colourMode: int = 0) -> str:
 	if len(findings) == 0:
 		return "No findings"
 
-	heading = heading if heading is not None else \
- "# Findings\nFind a list of findings below ordered by severity"
+	heading = (
+		heading
+		if heading is not None
+		else "# Findings\nFind a list of findings below ordered by severity"
+	)
 	strBuf = [heading]
 	findings = sorted(findings, key=lambda i: i["severity"], reverse=True)
 
@@ -77,17 +79,20 @@ colourMode: int = 0) -> str:
 
 	# Details
 	for finding in findings:
-		strBuf.extend([
-		f"## {finding['title']}", f"{finding['description']}",
-		f"\n\nFile: `{finding['file']}`",
-		f"### Severity\n\n{finding['severity']} (confidence: {finding['confidence']})",
-		f"### Evidence\n\nLine: {finding['line']}\n",
-		f"```python\n{formatEvidence(finding['evidence'])}\n```", ])
+		strBuf.extend(
+			[
+				f"## {finding['title']}",
+				f"{finding['description']}",
+				f"\n\nFile: `{finding['file']}`",
+				f"### Severity\n\n{finding['severity']} (confidence: {finding['confidence']})",
+				f"### Evidence\n\nLine: {finding['line']}\n",
+				f"```python\n{formatEvidence(finding['evidence'])}\n```",
+			]
+		)
 	return "\n".join(strBuf) + "\n"
 
 
-def json(findings: list[Finding], heading: Optional[str] = None,
-colourMode: int = 0) -> str:
+def json(findings: list[Finding], heading: Optional[str] = None, colourMode: int = 0) -> str:
 	"""Format to Json.
 
 	Args:
@@ -98,14 +103,16 @@ colourMode: int = 0) -> str:
 		str: String to write to a file of stdout
 	"""
 	findings = sorted(findings, key=lambda i: i["severity"], reverse=True)
-	out = {"heading": heading if heading is not None else \
- "Findings - Findings below are ordered by severity",
-	"findings": findings}
+	out = {
+		"heading": heading
+		if heading is not None
+		else "Findings - Findings below are ordered by severity",
+		"findings": findings,
+	}
 	return dumps(out, indent="\t")
 
 
-def csv(findings: list[Finding], heading: Optional[str] = None,
-colourMode: int = 0) -> str:
+def csv(findings: list[Finding], heading: Optional[str] = None, colourMode: int = 0) -> str:
 	"""Format to CSV.
 
 	Args:
@@ -118,21 +125,33 @@ colourMode: int = 0) -> str:
 	findings = sorted(findings, key=lambda i: i["severity"], reverse=True)
 	output = StringIO()
 	csvString = writer(output)
-	csvString.writerow([heading if heading is not None else \
- "Findings - Findings below are ordered by severity (you may want to delete this line)"])
-	csvString.writerow([
-	"id", "title", "description", "file", "evidence", "severity", "confidence",
-	"line"])
+	csvString.writerow(
+		[
+			heading
+			if heading is not None
+			else "Findings - Findings below are ordered by severity (you may want to delete this line)"
+		]
+	)
+	csvString.writerow(
+		["id", "title", "description", "file", "evidence", "severity", "confidence", "line"]
+	)
 	for finding in findings:
-		csvString.writerow([
-		finding["id"], finding["title"], finding["description"], finding["file"],
-		formatEvidence(finding["evidence"], False), finding["severity"],
-		finding["confidence"], finding["line"]])
+		csvString.writerow(
+			[
+				finding["id"],
+				finding["title"],
+				finding["description"],
+				finding["file"],
+				formatEvidence(finding["evidence"], False),
+				finding["severity"],
+				finding["confidence"],
+				finding["line"],
+			]
+		)
 	return output.getvalue()
 
 
-def ansi(findings: list[Finding], heading: Optional[str] = None,
-colourMode: int = 0) -> str:
+def ansi(findings: list[Finding], heading: Optional[str] = None, colourMode: int = 0) -> str:
 	"""Format to ansi.
 
 	Args:
@@ -143,7 +162,8 @@ colourMode: int = 0) -> str:
 		str: String to write to a file of stdout
 	"""
 	# pylint: disable=invalid-name
-	FMT = {"TXT": "",
+	FMT = {
+		"TXT": "",
 		"BLD": "",
 		"CLS": "",
 		"UL": "",
@@ -151,9 +171,10 @@ colourMode: int = 0) -> str:
 		"CG": "",
 		"CY": "",
 		"CODE": "│",
-	} # yapf: disable
+	}
 	if colourMode == 1:
-		FMT = {"TXT": "",
+		FMT = {
+			"TXT": "",
 			"BLD": "\033[01m",
 			"CLS": "\033[00m",
 			"UL": "\033[04m",
@@ -161,9 +182,10 @@ colourMode: int = 0) -> str:
 			"CG": "\033[32m",
 			"CY": "\033[33m",
 			"CODE": "│\033[100m\033[93m",
-		} # yapf: disable
+		}
 	elif colourMode == 2:
-		FMT = {"TXT": "\033[97m",
+		FMT = {
+			"TXT": "\033[97m",
 			"BLD": "\033[01m",
 			"CLS": "\033[00m",
 			"UL": "\033[04m",
@@ -171,19 +193,26 @@ colourMode: int = 0) -> str:
 			"CG": "\033[92m",
 			"CY": "\033[93m",
 			"CODE": "\033[97m│\033[107m\033[90m",
-		} # yapf: disable
+		}
 	if len(findings) == 0:
 		return f"{FMT['BLD']}{FMT['UL']}{FMT['CB']}No findings{FMT['CLS']}"
 
 	# pylint: enable=invalid-name
-	strBuf = [heading] if heading is not None else [
-	f"{FMT['BLD']}{FMT['UL']}{FMT['CB']}Findings{FMT['CLS']}\n",
-	f"{FMT['TXT']}Find a list of findings below ordered by severity\n"]
+	strBuf = (
+		[heading]
+		if heading is not None
+		else [
+			f"{FMT['BLD']}{FMT['UL']}{FMT['CB']}Findings{FMT['CLS']}\n",
+			f"{FMT['TXT']}Find a list of findings below ordered by severity\n",
+		]
+	)
 	findings = sorted(findings, key=lambda i: i["severity"], reverse=True)
 
 	# Summary Table
 	strBuf.append(f"{FMT['TXT']}┌{'─'*10}┬{'─'*50}┐")
-	strBuf.append("│Severity  │Finding                                           │") # yapf: disable
+	strBuf.append(
+		"│Severity  │Finding                                           │"
+	)
 	strBuf.append(f"├{'─'*10}┼{'─'*50}┤")
 	for finding in findings:
 		strBuf.append(f"│{finding['severity']: <10}│{finding['title'][:50]: <50}│")
@@ -193,23 +222,27 @@ colourMode: int = 0) -> str:
 	# Details
 	for finding in findings:
 		evidence = [f"{FMT['TXT']}┌{' ' + finding['file'] + ' ':─^85}┐"]
-		for line in finding['evidence']:
-			evidence.append((FMT['CODE'] if line["selected"] else f"{FMT['TXT']}│")
-			+ f"{str(line['line'])[:3]: >3}  "
-			+ f"{line['content'][:80]: <80}{FMT['CLS']}{FMT['TXT']}│")
+		for line in finding["evidence"]:
+			evidence.append(
+				(FMT["CODE"] if line["selected"] else f"{FMT['TXT']}│")
+				+ f"{str(line['line'])[:3]: >3}  "
+				+ f"{line['content'][:80]: <80}{FMT['CLS']}{FMT['TXT']}│"
+			)
 		evidence.append(f"└{'─'*85}┘")
-		evidenceStr = '\n'.join(evidence)
-		strBuf.extend([
-		f"{FMT['BLD']}{FMT['UL']}{FMT['CG']}{finding['title']}{FMT['CLS']}",
-		f"{FMT['TXT']}{finding['description']}",
-		f"\n{FMT['UL']}{FMT['CY']}Severity: {finding['severity']} "
-		+ f"(confidence: {finding['confidence']}){FMT['CLS']}\n",
-		f"{FMT['UL']}{FMT['CY']}Evidence{FMT['CLS']}\n{evidenceStr}\n", ])
+		evidenceStr = "\n".join(evidence)
+		strBuf.extend(
+			[
+				f"{FMT['BLD']}{FMT['UL']}{FMT['CG']}{finding['title']}{FMT['CLS']}",
+				f"{FMT['TXT']}{finding['description']}",
+				f"\n{FMT['UL']}{FMT['CY']}Severity: {finding['severity']} "
+				+ f"(confidence: {finding['confidence']}){FMT['CLS']}\n",
+				f"{FMT['UL']}{FMT['CY']}Evidence{FMT['CLS']}\n{evidenceStr}\n",
+			]
+		)
 	return "\n".join(strBuf) + f"{FMT['CLS']}"
 
 
-def sarif(findings: list[Finding], heading: Optional[str] = None,
-colourMode: int = 0) -> str:
+def sarif(findings: list[Finding], heading: Optional[str] = None, colourMode: int = 0) -> str:
 	"""Format to sarif https://sarifweb.azurewebsites.net/.
 
 	Args:
@@ -220,30 +253,54 @@ colourMode: int = 0) -> str:
 		str: String to write to a file of stdout
 	"""
 	out = {
-	"version": "2.1.0",
-	"$schema": "https://raw.githubusercontent.com/oasis-tcs/sarif-spec/master/Schemata/sarif-schema-2.1.0.json",
-	"runs": [{
-	"tool": {"driver": {
-		"name": "SimpleSecurity",
-		"informationUri": "https://github.com/FHPythonUtils/SimpleSecurity",
-		"version": "2020.*", }},
-	"results": [{
-		"ruleId": finding["id"],
-		"level": finding["severity"].toSarif(),
-		"message": {"text": f"{finding['title']}: {finding['description']}"},
-		"locations": [{
-		"physicalLocation": {
-			"artifactLocation": {"uri": finding["file"]},
-			"region": {
-				"startLine": finding["line"],
-				"snippet": {
-				"text": "".join([line["content"] for line in finding["evidence"] if line["selected"]])
-				}},
-			"contextRegion": {
-				"startLine": finding["evidence"][0]["line"],
-				"endLine": finding["evidence"][-1]["line"],
-				"snippet": {"text": "\n".join([line["content"] for line in finding["evidence"]])}}
-		}}]
-	} for finding in findings]
-	}]} # yapf: disable
+		"version": "2.1.0",
+		"$schema": "https://raw.githubusercontent.com/oasis-tcs/sarif-spec/master/Schemata/sarif-schema-2.1.0.json",
+		"runs": [
+			{
+				"tool": {
+					"driver": {
+						"name": "SimpleSecurity",
+						"informationUri": "https://github.com/FHPythonUtils/SimpleSecurity",
+						"version": "2020.*",
+					}
+				},
+				"results": [
+					{
+						"ruleId": finding["id"],
+						"level": finding["severity"].toSarif(),
+						"message": {"text": f"{finding['title']}: {finding['description']}"},
+						"locations": [
+							{
+								"physicalLocation": {
+									"artifactLocation": {"uri": finding["file"]},
+									"region": {
+										"startLine": finding["line"],
+										"snippet": {
+											"text": "".join(
+												[
+													line["content"]
+													for line in finding["evidence"]
+													if line["selected"]
+												]
+											)
+										},
+									},
+									"contextRegion": {
+										"startLine": finding["evidence"][0]["line"],
+										"endLine": finding["evidence"][-1]["line"],
+										"snippet": {
+											"text": "\n".join(
+												[line["content"] for line in finding["evidence"]]
+											)
+										},
+									},
+								}
+							}
+						],
+					}
+					for finding in findings
+				],
+			}
+		],
+	}
 	return dumps(out, indent="\t")