+++ simplesecurity/level.py
@@ -16,6 +16,7 @@ class Level(IntEnum):
 	MED = 2
 	HIGH = 3
 	"""
+
 	UNKNOWN = 0
 	LOW = 1
 	MED = 2
@@ -36,13 +37,19 @@ class Level(IntEnum):
 			str: string representation of a level
 		"""
 		reprMap = {
-		Level.UNKNOWN: "Unknown", Level.LOW: "Low", Level.MED: "Medium",
-		Level.HIGH: "High"}
+			Level.UNKNOWN: "Unknown",
+			Level.LOW: "Low",
+			Level.MED: "Medium",
+			Level.HIGH: "High",
+		}
 		return reprMap[self]
 
 	def toSarif(self) -> str:
 		"""Convert to sarif representation."""
 		reprMap = {
-		Level.UNKNOWN: "note", Level.LOW: "warning", Level.MED: "warning",
-		Level.HIGH: "error"}
+			Level.UNKNOWN: "note",
+			Level.LOW: "warning",
+			Level.MED: "warning",
+			Level.HIGH: "error",
+		}
 		return reprMap[self]